Move auth related api calls from users.api to auth.api
